requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.31.1
streamlit==1.39.0
transformers==4.45.2
torch==2.4.1
pillow==10.4.0
pymilvus==2.4.7
python-multipart==0.0.12
requests==2.32.3
httpx==0.27.2
numpy==1.26.4
setuptools==70.0.0
marshmallow==3.21.3
